+++ packages/dt_for_itables/src/index.js
@@ -12,6 +12,7 @@ DataTable.Buttons.jszip(JSZip);
 
 import 'datatables.net-columncontrol-dt';
 import 'datatables.net-fixedcolumns-dt';
+import 'datatables.net-fixedheader-dt';
 import 'datatables.net-keytable-dt';
 import 'datatables.net-rowgroup-dt';
 import DateTime from 'datatables.net-datetime';