+++ packages/dt_for_itables/CHANGELOG.md
@@ -1,7 +1,10 @@
+# 2.5.1 (2026-02-22)
+
+- We have added the FixedHeader extension of DataTables to the `dt_for_itables` package ([#496](https://github.com/mwouts/itables/pull/496))
+
 # 2.5.0 (2026-02-01)
 
 - We have updated datatables and its extensions to `datatables.net-dt==2.3.7`
-
 
 # 2.4.0 (2025-08-23)